feat: page detail pokemon, turn button & sprites type
- Also fixed warnings in console like synchronous access to params: now awaited, this sould solve the issue
