+++ src/libs/poke/dto/common.ts
@@ -15,3 +15,7 @@ export interface ExternalResource {
     name: string; // name of element
     url: string;  // url to call to get the info related to this element
 };
+
+export interface Indexable {
+  [key: string]: any
+};
