+++ src/libs/poke/dto/pokemon.ts
@@ -1,8 +1,7 @@
-import { ExternalResource } from "./common";
+import { ExternalResource, Indexable } from "./common";
 
 export interface PokemonEvolution {
     id: number;
-    
 };
 
 export interface PokemonStat {
@@ -11,28 +10,36 @@ export interface PokemonStat {
     stat: ExternalResource;
 };
 
+export interface PokemonType {
+    slot: number;
+    type: ExternalResource;
+}
+
+export interface PokemonSprites extends Indexable {
+    back_default: string | null;
+    back_female: string | null;
+    back_shiny: string | null;
+    back_shiny_female: string | null;
+    front_default: string | null;
+    front_female: string | null;
+    front_shiny: string | null;
+    front_shiny_female: string | null;
+};
+
 export interface Pokemon {
     id: number;
     name: string;
     order: number;
     species: ExternalResource;
-    sprites: {
-        back_default: string | null;
-        back_female: string | null;
-        back_shiny: string | null;
-        back_shiny_female: string | null;
-        front_default: string | null;
-        front_female: string | null;
-        front_shiny: string | null;
-        front_shiny_female: string | null;
-    };
+    sprites: PokemonSprites;
     stats: {
         base_stat: number;
         effort: number;
         stat: ExternalResource;
     }[];
-    types: {
-        slot: number;
-        type: ExternalResource;
-    }
+    types: PokemonType[];
+    past_types: {
+        generation: ExternalResource;
+        types: PokemonType[];
+    }[];
 };
