+++ tsconfig.json
@@ -23,7 +23,7 @@
 
     /* Path aliases */
     "paths": {
-      "~/*": ["./src/*"]
+      "~/*": ["./src/*"],
     },
 
     "types": ["vitest/globals"]
